Fix: Move classes used in static analysis into different namespaces
