+++ tests/StaticAnalysis/Src/Bar.php
@@ -11,9 +11,9 @@ declare(strict_types=1);
  * @see https://github.com/Jan0707/phpstan-prophecy
  */
 
-namespace JanGregor\Prophecy\Test\StaticAnalysis;
+namespace JanGregor\Prophecy\Test\StaticAnalysis\Src;
 
-interface Foo
+interface Bar
 {
-    public function foo(): string;
+    public function bar(): string;
 }
