+++ tests/StaticAnalysis/Src/Baz.php
@@ -11,9 +11,9 @@ declare(strict_types=1);
  * @see https://github.com/Jan0707/phpstan-prophecy
  */
 
-namespace JanGregor\Prophecy\Test\StaticAnalysis;
+namespace JanGregor\Prophecy\Test\StaticAnalysis\Src;
 
-interface Bar
+abstract class Baz
 {
-    public function bar(): string;
+    abstract public function baz(): string;
 }
