+++ tests/StaticAnalysis/Src/Foo.php
@@ -11,9 +11,9 @@ declare(strict_types=1);
  * @see https://github.com/Jan0707/phpstan-prophecy
  */
 
-namespace JanGregor\Prophecy\Test\StaticAnalysis;
+namespace JanGregor\Prophecy\Test\StaticAnalysis\Src;
 
-abstract class Baz
+interface Foo
 {
-    abstract public function baz(): string;
+    public function foo(): string;
 }
